fix: Ticket Status
Set the right views ID for pending status. Create function to format tickets
