--- server_ticket_viewer/data.js
@@ -1,15 +1,0 @@
-// 20210506201844
-// https://felipesolutions.zendesk.com/api/v2/tickets.json?page[size]=25
-
-{
-
-    "meta": {
-        "has_more": true,
-            "after_cursor": "eyJvIjoibmljZV9pZCIsInYiOiJhUmtBQUFBQUFBQUEifQ==",
-                "before_cursor": "eyJvIjoibmljZV9pZCIsInYiOiJhUUVBQUFBQUFBQUEifQ=="
-    },
-    "links": {
-        "prev": "https://felipesolutions.zendesk.com/api/v2/tickets.json?page%5Bbefore%5D=eyJvIjoibmljZV9pZCIsInYiOiJhUUVBQUFBQUFBQUEifQ%3D%3D&page%5Bsize%5D=25",
-            "next": "https://felipesolutions.zendesk.com/api/v2/tickets.json?page%5Bafter%5D=eyJvIjoibmljZV9pZCIsInYiOiJhUmtBQUFBQUFBQUEifQ%3D%3D&page%5Bsize%5D=25"
-    }
-}
